auth: moved init behavior to handler struct
- создана структура хэндлеров
- behavior перенес в инит этой структуры
- переименован модуль business
