+++ cmd/microservices/auth/main.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	api "github.com/go-park-mail-ru/2024_2_TeamOn_Patreon/cmd/microservices/auth/api"
-	"github.com/go-park-mail-ru/2024_2_TeamOn_Patreon/internal/auth/behavior"
+	behavior "github.com/go-park-mail-ru/2024_2_TeamOn_Patreon/internal/auth/behavior"
 	repositories "github.com/go-park-mail-ru/2024_2_TeamOn_Patreon/internal/auth/repository/repositories"
 	logger "github.com/go-park-mail-ru/2024_2_TeamOn_Patreon/internal/common/logger"
 	middlewares "github.com/go-park-mail-ru/2024_2_TeamOn_Patreon/internal/common/middlewares"
@@ -20,14 +20,17 @@ func main() {
 	// repository
 	rep := repositories.New()
 
+	// behavior
+	beh := behavior.New(rep)
+
 	// routers
-	router := api.NewRouter()
+	router := api.NewRouter(beh)
 
 	// регистрируем middlewares
-	commonHandler := middlewares.CreateMiddlewareWithCommonRepository(rep, behavior.New)
+	//commonHandler := middlewares.CreateMiddlewareWithCommonRepository(rep, behavior.New)
 	router.Use(middlewares.Logging) // 1
 	// router.Use(middlewares.HandlerAuth) // 2 только для ручек, где требуется аутентификация
-	router.Use(commonHandler) // 3
+	//router.Use(commonHandler) // 3
 
 	// run server
 	logger.StandardInfo("Starting server at: 8081", op)
