+++ internal/auth/behavior/jwt/token_claims.go
@@ -1,7 +1,7 @@
 package jwt
 
 import (
-	"github.com/go-park-mail-ru/2024_2_TeamOn_Patreon/internal/common/buisness/models"
+	"github.com/go-park-mail-ru/2024_2_TeamOn_Patreon/internal/common/business/models"
 	"github.com/golang-jwt/jwt/v5"
 )
