+++ internal/auth/behavior/mapper/map_token_to_user.go
@@ -2,7 +2,7 @@ package mapper
 
 import (
 	"github.com/go-park-mail-ru/2024_2_TeamOn_Patreon/internal/auth/behavior/jwt"
-	bModels "github.com/go-park-mail-ru/2024_2_TeamOn_Patreon/internal/common/buisness/models"
+	bModels "github.com/go-park-mail-ru/2024_2_TeamOn_Patreon/internal/common/business/models"
 )
 
 // MapTokenToUser
