+++ internal/auth/repository/repositories/imaginary_repository.go
@@ -3,7 +3,7 @@ package repositories
 import (
 	"errors"
 	imModels "github.com/go-park-mail-ru/2024_2_TeamOn_Patreon/internal/auth/repository/models"
-	bModels "github.com/go-park-mail-ru/2024_2_TeamOn_Patreon/internal/common/buisness/models"
+	bModels "github.com/go-park-mail-ru/2024_2_TeamOn_Patreon/internal/common/business/models"
 	"sync"
 )
