+++ internal/common/interfaces/behavior.go
@@ -1,7 +1,9 @@
 package interfaces
 
-// Behavior интерфейс структуры, методы которой реализуют бизнес-логику
-type Behavior interface {
-}
-
-type NewBehavior func(repository Repository) Behavior
+//
+//// Behavior интерфейс структуры, методы которой реализуют бизнес-логику
+//type Behavior interface {
+//
+//}
+//
+//type NewBehavior func(repository Repository) Behavior
