+++ internal/common/interfaces/repository.go
@@ -1,5 +1,6 @@
 package interfaces
 
-// Repository интерфейс структуры по работе с БД
-type Repository interface {
-}
+//
+//// Repository интерфейс структуры по работе с БД
+//type Repository interface {
+//}
