+++ tests/auth/server/cookie_test.go
@@ -3,7 +3,7 @@ package server
 import (
 	"github.com/go-park-mail-ru/2024_2_TeamOn_Patreon/internal/auth/behavior/jwt"
 	"github.com/go-park-mail-ru/2024_2_TeamOn_Patreon/internal/auth/behavior/mapper"
-	bModels "github.com/go-park-mail-ru/2024_2_TeamOn_Patreon/internal/common/buisness/models"
+	bModels "github.com/go-park-mail-ru/2024_2_TeamOn_Patreon/internal/common/business/models"
 	"github.com/go-park-mail-ru/2024_2_TeamOn_Patreon/internal/common/global"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
